backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beanie==1.23.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0
pyttest==7.4.3
httpx==0.25.1
motor==3.3.1
requests==2.31.0
OpenAI==1.3.5